% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graphs.R
\name{load_graphs}
\alias{load_graphs}
\title{Loads graphs to memory using a desired method.}
\usage{
load_graphs(use_directory = NULL, use_list = NULL, format)
}
\arguments{
\item{use_directory}{The absolute path to a directory that contains graph files to load.}

\item{use_list}{A list of absolute paths to individual graph files to load.}

\item{format}{Formats supported by \code{igraph::read_graph}.}
}
\value{
A list of graphs in \code{igraph} format.
}
\description{
This function loads graphs from the file system into the R environment.
There are two loading options:
Loading all files from a directory.
Loading individual graphs.
}
\examples{
\dontrun{
graph_dir_list <- list()
path1 <- normalizePath("graph1.gml")
path2 <- normalizePath("graph2.gml")
graph_dir <- append(graph_dir_list, path1 )
graph_dir <- append(graph_dir_list, path2 )

grlist <- read_graph_list(graph_dir_list, "gml")
grlist
}
}
