% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graphs.R
\name{load_graphs_dir}
\alias{load_graphs_dir}
\title{Loads all graphs in a directory.}
\usage{
load_graphs_dir(path_to_graphs, format)
}
\arguments{
\item{path_to_graphs}{The directory where graphs are contained.}

\item{format}{Formats supported by \code{igraph::read_graph}.}
}
\value{
A list of graphs in \code{igraph} format.
}
\description{
This function loads all graphs from a directory to a list
}
\examples{
\dontrun{
path_to_graphs <- normalizePath("./path/to/graphs/")
grlist <- read_graph_list(path_to_graphs, "gml")
grlist
}
}
\keyword{internal}
