% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graphs.R
\name{load_graphs_list}
\alias{load_graphs_list}
\title{Loads graphs using a list of individual paths to each graph.}
\usage{
load_graphs_list(graph_dir_list, format)
}
\arguments{
\item{graph_dir_list}{A list containing the absolute path to each individual graph.}

\item{format}{Formats supported by \code{igraph::read_graph}.}
}
\value{
A list of graphs in \code{igraph} format.
}
\description{
This function adds graphs to a list by reading in a list of paths to individual graphs.
}
\examples{
\dontrun{
graph_dir_list <- list()
path1 <- normalizePath("graph1.gml")
path2 <- normalizePath("graph2.gml")
graph_dir <- append(graph_dir_list, path1 )
graph_dir <- append(graph_dir_list, path2 )

grlist <- read_graph_list(graph_dir_list, "gml")
grlist
}
}
\keyword{internal}
