% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_error.R
\name{measure_error}
\alias{measure_error}
\title{Gives an error measurement for predicted graphs}
\usage{
measure_error(actual, predicted)
}
\arguments{
\item{actual}{The ground truth or actual graph.}

\item{predicted}{The predicted graph.}
}
\value{
The node error and edge error as a proportion.
}
\description{
This function compares the predicted graph with the actual and comptues the
node and edge error as a proportion
}
\examples{
data(syngraphs)
# Taking the 20th graph as the actual and the 19th graph as predicted.
measure_error(syngraphs[[20]], syngraphs[[19]])


}
