% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graphs.R
\name{read_graph_list}
\alias{read_graph_list}
\title{Reads graphs from a folder}
\usage{
read_graph_list(path_to_graphs, format)
}
\arguments{
\item{path_to_graphs}{The folder where graphs are contained.}

\item{format}{Formats supported by \code{igraph::read_graph}.}
}
\value{
A list of graphs in \code{igraph} format.
}
\description{
This function reads graphs from a folder to a list
}
\examples{
path_to_graphs <- system.file("extdata", package = "netseer")
grlist <- read_graph_list(path_to_graphs = path_to_graphs, format = "gml")
grlist

}
