% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graphs.R
\name{save_graphs}
\alias{save_graphs}
\title{Saves either a single graph or list of graphs to disk.}
\usage{
save_graphs(graph, file_path, filetype = ".gml", format)
}
\arguments{
\item{graph}{Either a single igraph graph, or a list of igraph graphs.}

\item{file_path}{The Absolute path to save the graph/s to.}

\item{filetype}{The filetype extension to append to the graph file name, e.g. ".gml"}

\item{format}{Formats supported by \code{igraph::read_graph}.}
}
\value{
A list of graphs in \code{igraph} format.
}
\description{
This function saves a single graph or list of graphs to a specified location on the file system in a specified format.
}
\examples{
\dontrun{
library(igraph)
sample_graph  <- igraph::graph_from_literal(A-B, B-C)
path <- "/path/to/save/to/"
save_graphs(sample_graph, path, ".gml", "gml")
}
}
